project/requirements.txt
fastapi==0.109.0
pydantic-settings==2.1.0
uvicorn==0.26.0
asyncpg==0.29.0
tortoise-orm==0.20.0
aerich==0.7.2
httpx==0.26.0
pytest==7.4.4
gunicorn==21.0.1
pytest-cov==4.1.0
flake8==7.0.0
black==23.12.1
